Reset1 2차 수정버전 출첵 ui 오류 1차 수정
