+++ .idea/shelf/2023-05-08__8_10__________.xml
@@ -1,0 +1,9 @@
+<changelist name="2023-05-08_오후_8_10,_체크아웃_전에_커밋되지_않은_변경_내용_[변경]" date="1683544273511" recycled="true" deleted="true">
+  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/2023-05-08_오후_8_10,_체크아웃_전에_커밋되지_않은_변경_내용_[변경]/shelved.patch" />
+  <option name="DESCRIPTION" value="2023-05-08 오후 8:10, 체크아웃 전에 커밋되지 않은 변경 내용 [변경]" />
+  <binary>
+    <option name="BEFORE_PATH" value="app/src/main/assets/student_db.db" />
+    <option name="AFTER_PATH" value="app/src/main/assets/student_db.db" />
+    <option name="SHELVED_PATH" value="$PROJECT_DIR$/.idea/shelf/2023-05-08_오후_8_10,_체크아웃_전에_커밋되지_않은_변경_내용_[변경]/student_db.db" />
+  </binary>
+</changelist>
